requirements.txt
flask
whisper
torch
torchaudio
tqdm
numba
google-generativeai
smtplib
email-validator
